Everything is up and running now in the "Mixed GWR Leave One Out Uber Function.R".  There is an uber function in there too.
Calculating the SCV (the yhat after droping each observation) make megaMaker take about twice as long for models 2-8.  If we want to ignore scv for time concerns, the megaMaker function in my sandbox does not have the code to calculate those values.

The uber function only takes in sample size and error as parameters, so currently there is no way to change the true model.  I'm about to add a way to change that, where we give it the true model and there are a bunch of if statements with beta generators for that model.  This isn't the most elegant solution, but its what I have already.